# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/employees": {
      "get": {
        "operationId": "EmployeesController_getAllEmployees_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "sortField",
            "required": false,
            "in": "query",
            "schema": {
              "example": "createdAt",
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "example": "DESC",
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "example": "John Doe",
              "type": "string"
            }
          },
          {
            "name": "departmentId",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "example": "WORKING",
              "type": "string",
              "enum": [
                "WORKING",
                "RETIRED"
              ]
            }
          },
          {
            "name": "position",
            "required": false,
            "in": "query",
            "schema": {
              "example": "Software Engineer",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Employees"
        ]
      },
      "post": {
        "operationId": "EmployeesController_createEmployee_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmployeeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Employees"
        ]
      }
    },
...
    "/api/v1/employees/{id}/provision-account": {
      "post": {
        "operationId": "EmployeesController_provisionAccount_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProvisionAccountDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Employees"
        ]
      }
    },
    "/api/v1/employees/profile": {
      "patch": {
        "operationId": "EmployeesController_updateEmployeeProfile_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmployeeProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Employees"
        ]
      }
    },
    "/api/v1/employees/{id}": {
      "patch": {
        "operationId": "EmployeesController_updateEmployee_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmployeeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Employees"
        ]
      }
    },
    "/api/v1/employees/import": {
      "post": {
        "operationId": "EmployeesController_importEmployees_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Employees"
        ]
      }
    },
...
    "/api/v1/employee-histories": {
      "post": {
        "operationId": "EmployeeHistoriesController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmployeeHistoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "EmployeeHistories"
        ]
      },
      "get": {
        "operationId": "EmployeeHistoriesController_findAll_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "sortField",
            "required": false,
            "in": "query",
            "schema": {
              "example": "createdAt",
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "example": "DESC",
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "example": "john",
              "type": "string"
            }
          },
          {
            "name": "departmentId",
            "required": false,
            "in": "query",
            "schema": {
              "example": "department",
              "type": "number"
            }
          },
          {
            "name": "employeeId",
            "required": false,
            "in": "query",
            "schema": {
              "example": "employee",
              "type": "number"
            }
          },
          {
            "name": "startDateFrom",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "example": "2026-01-01",
              "type": "string"
            }
          },
          {
            "name": "startDateTo",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "example": "2026-07-01",
              "type": "string"
            }
          },
          {
            "name": "endDateFrom",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "example": "2026-12-31",
              "type": "string"
            }
          },
          {
            "name": "endDateTo",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "example": "2026-12-31",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "EmployeeHistories"
        ]
      }
    },
...
    "/api/v1/employee-histories/{id}": {
      "get": {
        "operationId": "EmployeeHistoriesController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "EmployeeHistories"
        ]
      },
      "patch": {
        "operationId": "EmployeeHistoriesController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmployeeHistoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "EmployeeHistories"
        ]
      },
      "delete": {
        "operationId": "EmployeeHistoriesController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "EmployeeHistories"
        ]
      }
    },
...
  },
  "info": {
    "title": "Human Resource Manager",
    "description": "Human Resource Manager API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "token": {
        "scheme": "Bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      }
    },
    "schemas": {
...
      "ProvisionAccountDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "test123@yopmail.com"
          },
          "password": {
            "type": "string",
            "example": "12345678!"
          },
          "role": {
            "type": "string",
            "example": "EMPLOYEE",
            "enum": [
              "ADMIN",
              "MANAGER",
              "EMPLOYEE"
            ]
          }
        },
        "required": [
          "email",
          "password",
          "role"
        ]
      },
      "CreateEmployeeDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "gender": {
            "type": "string",
            "example": "MALE",
            "enum": [
              "MALE",
              "FEMALE",
              "OTHER"
            ]
          },
          "birthday": {
            "format": "date-time",
            "type": "string",
            "example": "1990-01-01"
          },
          "phone": {
            "type": "string",
            "example": "0901234567"
          },
          "address": {
            "type": "string",
            "example": "Tien Du, Bac Ninh"
          },
          "hireDate": {
            "format": "date-time",
            "type": "string",
            "example": "2026-07-01"
          },
          "position": {
            "type": "string",
            "example": "Software Engineer"
          },
          "departmentId": {
            "type": "number",
            "example": 1
          },
          "status": {
            "type": "string",
            "example": "WORKING",
            "enum": [
              "WORKING",
              "RETIRED"
            ],
            "default": "WORKING"
          },
          "basicSalary": {
            "type": "number",
            "example": 10000000
          },
          "account": {
            "$ref": "#/components/schemas/ProvisionAccountDto"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "gender",
          "birthday",
          "phone",
          "address",
          "hireDate",
          "position",
          "departmentId",
          "status",
          "basicSalary",
          "account"
        ]
      },
      "UpdateEmployeeProfileDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "gender": {
            "type": "string",
            "example": "MALE",
            "enum": [
              "MALE",
              "FEMALE",
              "OTHER"
            ]
          },
          "phone": {
            "type": "string",
            "example": "0901234567"
          },
          "address": {
            "type": "string",
            "example": "Tien Du, Bac Ninh"
          },
          "birthday": {
            "format": "date-time",
            "type": "string",
            "example": "2026-07-01"
          },
          "avatar": {
            "type": "string",
            "example": "https://example.com/avatar.jpg"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "gender",
          "phone",
          "address",
          "birthday",
          "avatar"
        ]
      },
      "UpdateEmployeeDto": {
        "type": "object",
        "properties": {
          "hireDate": {
            "format": "date-time",
            "type": "string",
            "example": "2026-01-01"
          },
          "position": {
            "type": "string",
            "example": "Software Engineer"
          },
          "status": {
            "type": "string",
            "enum": [
              "WORKING",
              "RETIRED"
            ],
            "example": "WORKING"
          },
          "departmentId": {
            "type": "number",
            "example": "1"
          },
          "basicSalary": {
            "type": "number",
            "example": 10000000
          }
        },
        "required": [
          "hireDate",
          "position",
          "status",
          "departmentId",
          "basicSalary"
        ]
      },
...
      "CreateEmployeeHistoryDto": {
        "type": "object",
        "properties": {
          "employeeId": {
            "type": "number",
            "example": 1
          },
          "departmentId": {
            "type": "number",
            "example": 1
          },
          "position": {
            "type": "string",
            "example": "Software Engineer"
          },
          "startDate": {
            "format": "date-time",
            "type": "string",
            "example": "2026-01-01"
          },
          "endDate": {
            "format": "date-time",
            "type": "string",
            "example": "2026-01-01"
          },
          "basicSalary": {
            "type": "number",
            "example": 10000000
          }
        },
        "required": [
          "employeeId",
          "departmentId",
          "position",
          "startDate",
          "endDate",
          "basicSalary"
        ]
      },
      "UpdateEmployeeHistoryDto": {
        "type": "object",
        "properties": {
          "departmentId": {
            "type": "number",
            "example": 1
          },
          "position": {
            "type": "string",
            "example": "Software Engineer"
          },
          "startDate": {
            "format": "date-time",
            "type": "string",
            "example": "2026-01-01"
          },
          "endDate": {
            "format": "date-time",
            "type": "string",
            "example": "2026-01-01"
          }
        },
        "required": [
          "departmentId",
          "position",
          "startDate",
          "endDate"
        ]
      },
...
    }
  },
  "security": [
    {
      "token": []
    }
  ]
}